Similar test for nil

+++ Tests/MapboxMapsTests/MapView/Integration Tests/OfflineManagerIntegrationTests.swift
@@ -62,7 +62,9 @@ internal class OfflineManagerIntegrationTestCase: IntegrationTestCase {
         let expectation = self.expectation(description: "Wait...")
         _ = XCTWaiter.wait(for: [expectation], timeout: 1.0)
 
-        try TileStore.removeDirectory(at: tileStorePathURL)
+        if let tileStorePathURL = tileStorePathURL {
+            try TileStore.removeDirectory(at: tileStorePathURL)
+        }
     }
 
     // MARK: Test Cases
